Rename the 'cfg' package to 'config'
