+++ cmd/knoxite/main.go
@@ -18,7 +18,7 @@ import (
 	shutdown "github.com/klauspost/shutdown2"
 	"github.com/spf13/cobra"
 
-	"github.com/knoxite/knoxite/cfg"
+	"github.com/knoxite/knoxite/config"
 	_ "github.com/knoxite/knoxite/storage/azure"
 	_ "github.com/knoxite/knoxite/storage/backblaze"
 	_ "github.com/knoxite/knoxite/storage/dropbox"
@@ -43,7 +43,7 @@ var (
 	CommitSHA = ""
 
 	globalOpts = GlobalOptions{}
-	config     = &cfg.Config{}
+	cfg        = &config.Config{}
 
 	// RootCmd is the core command used for cli-arg parsing
 	RootCmd = &cobra.Command{
@@ -64,7 +64,7 @@ func main() {
 
 	RootCmd.PersistentFlags().StringVarP(&globalOpts.Repo, "repo", "r", "", "Repository directory to backup to/restore from (default: current working dir)")
 	RootCmd.PersistentFlags().StringVarP(&globalOpts.Password, "password", "p", "", "Password to use for data encryption")
-	RootCmd.PersistentFlags().StringVarP(&globalOpts.ConfigURL, "configURL", "C", cfg.DefaultPath(), "Path to the configuration file")
+	RootCmd.PersistentFlags().StringVarP(&globalOpts.ConfigURL, "configURL", "C", config.DefaultPath(), "Path to the configuration file")
 
 	globalOpts.Repo = os.Getenv("KNOXITE_REPOSITORY")
 	globalOpts.Password = os.Getenv("KNOXITE_PASSWORD")
@@ -93,12 +93,12 @@ func init() {
 // ConfigURL flag.
 func initConfig() {
 	var err error
-	config, err = cfg.New(globalOpts.ConfigURL)
+	cfg, err = config.New(globalOpts.ConfigURL)
 	if err != nil {
 		log.Fatalf("error reading the config file: %v\n", err)
 		return
 	}
-	if err = config.Load(); err != nil {
+	if err = cfg.Load(); err != nil {
 		log.Fatalf("error loading the config file: %v\n", err)
 		return
 	}
@@ -106,7 +106,7 @@ func initConfig() {
 	// There can occur a panic due to an entry assigment in nil map when theres
 	// no map initialized to store the RepoConfigs. This will prevent this from
 	// happening:
-	if config.Repositories == nil {
-		config.Repositories = make(map[string]cfg.RepoConfig)
+	if cfg.Repositories == nil {
+		cfg.Repositories = make(map[string]config.RepoConfig)
 	}
 }
