+++ cmd/knoxite/repository.go
@@ -228,7 +228,7 @@ func openRepository(path, password string) (knoxite.Repository, error) {
 		}
 	}
 
-	if rep, ok := config.Repositories[path]; ok {
+	if rep, ok := cfg.Repositories[path]; ok {
 		return knoxite.OpenRepository(rep.Url, password)
 	}
 	return knoxite.OpenRepository(path, password)
